+++ example/lib/main.dart
@@ -29,16 +29,15 @@ class MyHomePage extends StatelessWidget {
         title: Text('DelayedWidget Example'),
       ),
       body: Center(
-          child: DelayedWidget(            
-            delayDuration: Duration(milliseconds: 200),// Not required
-            animationDuration: Duration(seconds: 1),// Not required
-              animation: DelayedAnimations.SLIDE_FROM_BOTTOM,// Not required
+          child: DelayedWidget(
+              delayDuration: Duration(milliseconds: 200), // Not required
+              animationDuration: Duration(seconds: 1), // Not required
+              animation: DelayedAnimations.SLIDE_FROM_BOTTOM, // Not required
               child: Container(
                 width: 200,
                 height: 200,
                 color: Colors.red,
               ))),
-      
     );
   }
 }