+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 ## [1.0.1] - 2020-12-25.
 
 * Updates readme.
+
+## [1.0.2] - 2021-02-09.
+
+* Changes default animation.
+* Adds more examples to Example Project.
+* Adds demo gifs.